+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/graph/DependencyGraphNode.java
@@ -29,7 +29,7 @@ public interface DependencyGraphNode {
      * Returns a simple id for this node, unique across all nodes in the same graph.
      * This id cannot be used across graphs.
      */
-    Long getNodeId();
+    Integer getNodeId();
 
     ResolvedConfigurationIdentifier getResolvedConfigurationId();
